[ALPS02248355] M4U: Support 4GB Dram
[Detail]
Some build error when CONFIG_MTK_LM_MODE on.
[Solution]
Cast to uintptr_t before pointer.
[Feature] Others

MTK-Commit-Id: c9a1ad6858b3051a45ad471d9473be81f3cbad89

CR-Id: ALPS02248355

+++ drivers/misc/mediatek/m4u/2.0/m4u_pgtable.c
@@ -434,7 +434,8 @@ int m4u_alloc_pte(m4u_domain_t *domain, imu_pgd_t *pgd, unsigned int pgprot)
 
 	/* check pte alignment -- must 1K align */
 	if (unlikely(pte_new & (IMU_BYTES_PER_PTE - 1))) {
-		m4u_aee_print("%s: fail, not algin pa=0x%p, va=0x%p\n", __func__, (void *)pte_new, pte_new_va);
+		m4u_aee_print("%s: fail, not algin pa=0x%p, va=0x%p\n", __func__,
+			(void *)(uintptr_t)pte_new, pte_new_va);
 		/* kfree(pte_new_va); */
 		kmem_cache_free(gM4u_pte_kmem, (void *)pte_new_va);
 		return -ENOMEM;
